// src/app/Model/companymodel.ts
export interface companymodel {
  id: number;
  srno: number;
  HistoryId: number;
  historyId: number;
  vamid: number;
  resourceName: string;
  manager: string;
  Delaydays: number;
  email: string;
  programName: string;
  TechTrack: string;
  category: string;
  startDate: Date;
  endDate: Date;
  SMEName: string;
  programStatus: string;
  SMEStatus: string;
  programsTrackerId: number;
  smecomments: string;
  programcode: string;
}